requirements.txt

fastapi
uvicorn
tensorflow
pillow
python-multipart
numpy
gdown

